Make the library compatible with the Arduino IDE
Add library.properties metadata file. This file is required for libraries with the 1.5 folder structure (source files under the src subfolder) in order for the Arduino IDE to recognize the library.

Move example sketch to appropriately named folder. This change causes the example sketch to be accessible via the Arduino IDE's File > Examples > ESP32-TFT-Library-ILI9486 menu after the library is installed.